Create Filter a long JSON file
In this program, we start by using the file_get_contents function to read in a JSON file called quotes.json and decoding it into a PHP object called $quotes.

Next, we use the array_filter function to filter the quotes that have "Buddha" as the author. This creates a new array of quotes called $buddha_quotes.

Finally, we use the json_encode function to encode the filtered quotes into a new JSON string called $buddha_quotes_json, and then use the file_put_contents function to write the string to a new JSON file called buddha_quotes.json.

Note that this program assumes that the JSON file is a list of objects, where each object has a key called "author" and a key called "quote". If your JSON file has a different structure, you may need to modify the program accordingly.