Create FilterJson.py
In this program, we start by using the open function to read in a JSON file called quotes.json and load it into a Python object called quotes.

Next, we use a list comprehension to filter the quotes that have "Buddha" as the author. This creates a new list of quotes called buddha_quotes.

Finally, we use the open function again to write the filtered quotes to a new JSON file called buddha_quotes.json using the dump function from the json package.

Note that this program assumes that the JSON file is a list of objects, where each object has a key called "author" and a key called "quote". If your JSON file has a different structure, you may need to modify the program accordingly.